% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{compileRInside}
\alias{compileRInside}
\title{Compile RInside D code}
\usage{
compileRInside(fileName, cleanup = TRUE)
}
\arguments{
\item{cleanup}{logical whether to remove the generated directories afterwards or not}

\item{fileNames}{single string item for path to file to be}
}
\description{
Simple function to compile a D script that calls the R API
}
