% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{create_file_name}
\alias{create_file_name}
\title{Creates a random file name}
\usage{
create_file_name(prefix = "script", extn = NULL, nrand = 12)
}
\arguments{
\item{prefix}{the file prefix}

\item{extn}{the file extension with or without a leading full stop}

\item{nrand}{the number of random characters to generate as an 
append to the file name}
}
\value{
A random file name
}
\description{
Creates a random file name
}
