% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{dir.copy}
\alias{dir.copy}
\title{unix-style function to copy a directory from}
\usage{
dir.copy(from, to, flags = "-r")
}
\arguments{
\item{from}{the path to the directory to be copied}

\item{to}{the path that the direectory should be copied to}

\item{flags}{character vector of the flags passed to the 
cp linux function. Defaults to "-r".}
}
\value{
invisibly returns the command that was run
}
\description{
unix-style function to copy a directory from
}
