% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{dir.remove}
\alias{dir.remove}
\title{Delete a directory}
\usage{
dir.remove(directory, flags = "-rf")
}
\arguments{
\item{directory}{character single directory to be removed}

\item{flags}{character vector of flags to be passed to the "rm"
unix function.}
}
\description{
Delete a directory
}
