% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{doLogging}
\alias{doLogging}
\title{Function sets option for logging}
\usage{
doLogging(flag)
}
\arguments{
\item{flag}{for whether functions should be verbose 
and print progress information}
}
\value{
the input flag invisibly
}
\description{
Function sets option for logging
}
