% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{.sauce}
\alias{.sauce}
\title{Workhorse function for compiling the D code and creating the R wrapper}
\usage{
.sauce(
  fileNames,
  dropFolder = NULL,
  folderName = NULL,
  compiler = "ldmd2",
  flags = c("-O", "-boundscheck=off", "-mcpu=native")
)
}
\arguments{
\item{fileNames}{character vector containing the module (file) names to 
be compiled and loaded into the R session. The file name must 
include the ".d" extension.}

\item{dropFolder}{logical for whether the folder containing the code
and compilation artifacts should be dropped after compilation is 
done. Defaults to TRUE.}

\item{folderName}{character for what the folder name the generated code
should be. Defaults to NULL - random name generated.}

\item{compiler}{character either "ldmd2" (default - ldc compiler) 
or "dmd" (DMD compiler)}

\item{flags}{character vector flags to be passed to the chosen D compiler}

\item{code}{character vector containing the code to be compiled}
}
\description{
The first file given is the main file containing functions
       to be exported, other files given are supporting files
}
