% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{extractModuleNames}
\alias{extractModuleNames}
\title{Extract name of file from extension}
\usage{
extractModuleNames(fileNames, extn = "d")
}
\arguments{
\item{fileNames}{character vector containing the file names}

\item{extn}{the extension to check for (with or without 
the starting full stop)}
}
\value{
character vector with item extracted
}
\description{
Extract name of file from extension
}
