% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{getExternalPtr}
\alias{getExternalPtr}
\title{Function to return pointer of an exported item from 
        loaded DLL table}
\usage{
getExternalPtr(symbolName, packageName = NULL)
}
\arguments{
\item{symbolName}{character for the name of the native 
symbol to be retrieved}

\item{packageName}{character the name of the package
can be left as NULL}
}
\value{
An object of class externalptr
}
\description{
Function to return pointer of an exported item from 
        loaded DLL table
}
\seealso{
getNativeSymbolInfo, getLoadedDLLs, dyn.load, 
             getDLLRegisteredRoutines
}
