% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{getSaucerOptions}
\alias{getSaucerOptions}
\title{Gets the specified compiler options}
\usage{
getSaucerOptions(names)
}
\arguments{
\item{names}{character vector for the names in .options
whose values are to be returned}
}
\value{
list containing values indexed with names
}
\description{
Gets the specified compiler options
}
