% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{hasExtension}
\alias{hasExtension}
\title{Function checks file names fileNames for whether 
       they contain the file extension extn}
\usage{
hasExtension(fileNames, extn = "d")
}
\arguments{
\item{fileNames}{character vector containing the file names}

\item{extn}{the extension to check for (with or without 
the begining full stop)}
}
\value{
logical vector for whether the files given by
        fileNames have the given extension
}
\description{
Function checks file names fileNames for whether 
       they contain the file extension extn
}
