% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{runCommand}
\alias{runCommand}
\title{Function to run a system command (usually compilation)}
\usage{
runCommand(command, runDirectory, failDirectory, successDirectory)
}
\arguments{
\item{command}{the command string to run}

\item{runDirectory}{the directory that the command should be run}

\item{failDirectory}{the directory that should be switched to if the command fails}

\item{successDirectory}{the directory that should be switched to in the event of a success}
}
\value{
invisibly returns NULL
}
\description{
Function to run a system command (usually compilation)
}
