% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{runDTests}
\alias{runDTests}
\title{Runs the internal D tests to make sure that the package is running properly}
\usage{
runDTests(dropFolder = TRUE, folderNamePrefix = "dTest")
}
\value{
NULL invisibly
}
\description{
Runs the internal D tests to make sure that the package is running properly
}
