% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{runDTranslatorTests}
\alias{runDTranslatorTests}
\title{Runs the internal D unittests for translator.d}
\usage{
runDTranslatorTests(dropFolder = TRUE, folderNamePrefix = "dTest")
}
\arguments{
\item{dropFolder}{whether the generated code folder should be dropped or not}

\item{folderNamePrefix}{the prefix name for the folder defaults to "dSaucerTest"}
}
\value{
NULL invisibly
}
\description{
Runs the internal D unittests for translator.d
}
