% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{sauce}
\alias{sauce}
\title{Function compiles and loads D modules into R session}
\usage{
sauce(fileNames, ...)
}
\arguments{
\item{fileNames}{character vector containing the module names to 
be compiled and loaded into the R session. Includes the
".d" extension (for now)
TODO: this functionality requires more work}

\item{folderName}{a character containing the name of the
folder where the code will be created. If missing
 a random name is generated}

\item{compiler}{character either "ldmd2" (default - ldc compiler) 
or "dmd" (DMD compiler)}

\item{dropFolder}{whether to delete the folder containing
the compilation artifacts once the compilation is done.}
}
\description{
The first file given is the main file containing functions
       to be exported, other files given are supporting files
}
\examples{

# require(saucer)
# sauce("script.d", dropFolder = TRUE)
# 
# x = seq(1.0, 10.0, by = 0.5); y = seq(1.0, 10.0, by = 0.5)
# generate_numbers(as.integer(100))
# dot(x, y)
# vdmul(x, y)
# ans1 = outer_prod_serial(x, y)
# ans2 = outer_prod_parallel(x, y) # parallel version
# sum(abs(ans1 - ans2)) == 0
# sexp_check()

}
