% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\docType{package}
\name{saucer-package}
\alias{saucer}
\alias{saucer-package}
\title{saucer: Package to call D code from R and vice versa}
\description{
Package that allows the user to write D code and call it from R and to call R code from D
}
\seealso{
To use this package see the \code{\link{dfunctions}}
             or the \code{\link{dfunctions}} function
}
\author{
\strong{Maintainer}: Chibisi Chima-Okereke \email{chibco@gmail.com}

}
\keyword{internal}
