% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{setLogFile}
\alias{setLogFile}
\title{Sets the file name where the log file will be written}
\usage{
setLogFile(fileName = NULL)
}
\arguments{
\item{fileName}{character for the file name where the log 
file will be written.}
}
\description{
Sets the file name where the log file will be written
}
