% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{setSaucerOptions}
\alias{setSaucerOptions}
\title{Sets the saucer compiler options}
\usage{
setSaucerOptions(...)
}
\arguments{
\item{...}{named parameters used to set
the options. See details.}
}
\description{
Sets the saucer compiler options
}
\details{
Options are as follows:

        compiler character the compiler to use, choose
                 "ldmd2", "dmd", other options will mean that
                 the flags will need to be reset
        DFlagsScript character array containing the flags
                 to be set for compiling the scripts
                 defaults to c("-c", "-g", "-J=\".\"")
        DTranslatorFlags character array containing the flags
                 to be set to generate the function wrappers
                 defaults to c("-J=\".\"")
        DOptionalFlags character array containing the flags to
                 be set for optional D flags e.g. optimisation etc.
                 defaults to c("-O", "-boundscheck=off", "-mcpu=native")
        RFlags character array for R flags for compiling e.g. -L-lR etc.
                 defaults to c("-fPIC", "-L-lR")
        SaucedTests compiler flags for the sauce tests
}
