% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{compileRInside}
\alias{compileRInside}
\title{Compile RInside D code}
\usage{
compileRInside(
  fileName,
  flags = c("-O", "-fPIC", "-L-lR", "-L-lRmath", "-mcpu=native"),
  cleanup = TRUE
)
}
\arguments{
\item{fileNames}{single string item for path to file to be}

\item{extraFiles}{single character string containing the extra files 
to be included in the compilation}
}
\description{
Simple function to compile a D script that calls the R API
}
