% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{dfunctions}
\alias{dfunctions}
\title{Compile D code from string}
\usage{
dfunctions(codeArr, dropFolder = TRUE, folderName = NULL, moduleName = NULL)
}
\arguments{
\item{code}{character containing the code to be compiled}
}
\value{
character array of code submitted into the code array
}
\description{
Compiles and loads D code from string array, note that the desired 
             export functions must still be marked with `@Export()`
}
\examples{
        dfunctions("
        import std.stdio: writeln;
        @Export() int squared(int x)
        {
          x = x*x;
          writeln(\"x squared: \", x);
          return x;
        }
        @Export() double cubed(double x)
        {
          x = x*x*x;
          writeln(\"x cubed: \", x);
          return x;
        }
        ")

        squared(as.integer(5))
        cubed(5)

}
