% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saucer.r
\name{sauce}
\alias{sauce}
\title{Function compiles and loads D modules into R session}
\usage{
sauce(fileNames, ...)
}
\arguments{
\item{fileNames}{character vector containing the module (file) names to 
be compiled and loaded into the R session. The file name must 
include the ".d" extension. Only functions in the first
file can be exported. Subsequent files are for dependencies
in the first file only.}

\item{code}{character vector containing the code to be compiled}

\item{dropFolder}{logical for whether the folder containing the code
and compilation artifacts should be dropped after compilation is 
done. Defaults to TRUE.}

\item{folderName}{character for what the folder name the generated code
should be. Defaults to NULL - random name generated.}

\item{compiler}{character either "ldmd2" (default - ldc compiler) 
or "dmd" (DMD compiler)}

\item{flags}{character vector flags to be passed to the chosen D compiler}
}
\description{
The first file given is the main file containing functions
       to be exported, other files given are supporting files
}
